Implemented fullscreen toggling
This was actually a larger change than the simple commit message implies. It turns out you can't change the decoration of a frame if it is "displayable", and as far as I know there is no way to make a displayable frame undisplayable.  This requires moving the JPanel from the old frame to a new, properly configured frame.

EDIT: so apparently dispose() makes frames undisplayable, after which I could change crap and then do a setVisible(true). Well, next time I'll know that.